fix memory leaks when parser throw error and add check nullable in setnode

+++ src/libs/Math.hpp
@@ -4,24 +4,27 @@
 #include <cmath>
 #include "../Debugger.hpp"
 
-namespace AutoLang {
-namespace Libs {
-namespace Math {
+namespace AutoLang
+{
+    namespace Libs
+    {
+        namespace Math
+        {
 
-void init(CompiledProgram& compile);
-long long integer_pow(long long base, long long exp);
-AObject* abs(NativeFuncInData);
-AObject* pow(NativeFuncInData);
-AObject* round(NativeFuncInData);
-AObject* floor(NativeFuncInData);
-AObject* ceil(NativeFuncInData);
-AObject* trunc(NativeFuncInData);
-AObject* fmod(NativeFuncInData);
-AObject* sin(NativeFuncInData);
-AObject* cos(NativeFuncInData);
-AObject* tan(NativeFuncInData);
+            void init(CompiledProgram &compile);
+            int64_t integer_pow(int64_t base, int64_t exp);
+            AObject *abs(NativeFuncInData);
+            AObject *pow(NativeFuncInData);
+            AObject *round(NativeFuncInData);
+            AObject *floor(NativeFuncInData);
+            AObject *ceil(NativeFuncInData);
+            AObject *trunc(NativeFuncInData);
+            AObject *fmod(NativeFuncInData);
+            AObject *sin(NativeFuncInData);
+            AObject *cos(NativeFuncInData);
+            AObject *tan(NativeFuncInData);
 
-}
-}
+        }
+    }
 }
 #endif
